fix: 중복된 config파일 통합
