--- src/main/java/com/kernel/sense_log/common/config/WebConfig.java
@@ -1,31 +1,0 @@
-package com.kernel.sense_log.common.config;
-
-import com.kernel.sense_log.domain.auth.LoginUserArgumentResolver;
-import com.kernel.sense_log.domain.auth.interceptor.AuthenticationInterceptor;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-
-    private final LoginUserArgumentResolver loginUserArgumentResolver;
-    private final AuthenticationInterceptor authenticationInterceptor;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(loginUserArgumentResolver);
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(authenticationInterceptor)
-                .order(1)
-                .addPathPatterns("/**")
-                .excludePathPatterns("/", "/auth/join", "/auth/login");
-    }
-}
